refactor: 뷰 모드 결정 로직을 별도의 유틸리티 함수로 분리
